+++ obj/src/Component.js
@@ -4,6 +4,7 @@ exports.Component = void 0;
 const pip_services3_commons_nodex_1 = require("pip-services3-commons-nodex");
 const CompositeLogger_1 = require("./log/CompositeLogger");
 const CompositeCounters_1 = require("./count/CompositeCounters");
+const CompositeTracer_1 = require("./trace/CompositeTracer");
 /**
  * Abstract component that supportes configurable dependencies, logging
  * and performance counters.
@@ -19,6 +20,7 @@ const CompositeCounters_1 = require("./count/CompositeCounters");
  *
  * - <code>\*:counters:\*:\*:1.0</code>     (optional) [[ICounters]] components to pass collected measurements
  * - <code>\*:logger:\*:\*:1.0</code>       (optional) [[ILogger]] components to pass log messages
+ * - <code>\*:tracer:\*:\*:1.0</code>       (optional) [[ITracer]] components to trace executed operations
  * - ...                                    References must match configured dependencies.
  */
 class Component {
@@ -26,6 +28,7 @@ class Component {
         this._dependencyResolver = new pip_services3_commons_nodex_1.DependencyResolver();
         this._logger = new CompositeLogger_1.CompositeLogger();
         this._counters = new CompositeCounters_1.CompositeCounters();
+        this._tracer = new CompositeTracer_1.CompositeTracer();
     }
     /**
      * Configures component by passing configuration parameters.
@@ -45,6 +48,7 @@ class Component {
         this._dependencyResolver.setReferences(references);
         this._logger.setReferences(references);
         this._counters.setReferences(references);
+        this._tracer.setReferences(references);
     }
 }
 exports.Component = Component;