+++ obj/src/count/ICounterTimingCallback.d.ts
@@ -2,16 +2,16 @@
 /**
  * Interface for a callback to end measurement of execution elapsed time.
  *
- * @see [[Timing]]
+ * @see [[CounterTiming]]
  */
-export interface ITimingCallback {
+export interface ICounterTimingCallback {
     /**
      * Ends measurement of execution elapsed time and updates specified counter.
      *
      * @param name      a counter name
      * @param elapsed   execution elapsed time in milliseconds to update the counter.
      *
-     * @see [[Timing.endTiming]]
+     * @see [[CounterTiming.endTiming]]
      */
     endTiming(name: string, elapsed: number): void;
 }