+++ obj/src/trace/ITracer.js
@@ -1,4 +1,4 @@
 "use strict";
-/** @module count */
+/** @module trace */
 Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=ITimingCallback.js.map
+//# sourceMappingURL=ITracer.js.map