+++ obj/src/trace/index.d.ts
@@ -1,0 +1,14 @@
+/**
+ * @module trace
+ *
+ * Todo: Rewrite the description
+ *
+ * @preferred
+ */
+export { ITracer } from './ITracer';
+export { NullTracer } from './NullTracer';
+export { CachedTracer } from './CachedTracer';
+export { LogTracer } from './LogTracer';
+export { CompositeTracer } from './CompositeTracer';
+export { OperationTrace } from './OperationTrace';
+export { DefaultTracerFactory } from './DefaultTracerFactory';