+++ src/Component.ts
@@ -10,6 +10,7 @@ import { IReferenceable } from 'pip-services3-commons-nodex';
 import { DependencyResolver } from 'pip-services3-commons-nodex';
 import { CompositeLogger } from './log/CompositeLogger';
 import { CompositeCounters } from './count/CompositeCounters';
+import { CompositeTracer } from './trace/CompositeTracer';
 
 /**
  * Abstract component that supportes configurable dependencies, logging
@@ -26,12 +27,14 @@ import { CompositeCounters } from './count/CompositeCounters';
  * 
  * - <code>\*:counters:\*:\*:1.0</code>     (optional) [[ICounters]] components to pass collected measurements
  * - <code>\*:logger:\*:\*:1.0</code>       (optional) [[ILogger]] components to pass log messages
+ * - <code>\*:tracer:\*:\*:1.0</code>       (optional) [[ITracer]] components to trace executed operations
  * - ...                                    References must match configured dependencies.
  */
 export class Component implements IConfigurable, IReferenceable {
     protected _dependencyResolver: DependencyResolver = new DependencyResolver();
     protected _logger: CompositeLogger = new CompositeLogger();
     protected _counters: CompositeCounters = new CompositeCounters();
+    protected _tracer: CompositeTracer = new CompositeTracer();
 
     /**
      * Configures component by passing configuration parameters.
@@ -52,5 +55,6 @@ export class Component implements IConfigurable, IReferenceable {
         this._dependencyResolver.setReferences(references);
         this._logger.setReferences(references);
         this._counters.setReferences(references);
+        this._tracer.setReferences(references);
     }
 }