+++ src/count/CachedCounters.ts
@@ -3,8 +3,8 @@ import { IReconfigurable } from 'pip-services3-commons-nodex';
 import { ConfigParams } from 'pip-services3-commons-nodex';
 
 import { ICounters } from './ICounters';
-import { Timing } from './Timing';
-import { ITimingCallback } from './ITimingCallback';
+import { CounterTiming } from './CounterTiming';
+import { ICounterTimingCallback } from './ICounterTimingCallback';
 import { CounterType } from './CounterType';
 import { Counter } from './Counter';
 
@@ -20,7 +20,7 @@ import { Counter } from './Counter';
  *     - reset_timeout:   timeout in milliseconds to reset the counters. 0 disables the reset 
  *     (default: 0)
  */
-export abstract class CachedCounters implements ICounters, IReconfigurable, ITimingCallback {
+export abstract class CachedCounters implements ICounters, IReconfigurable, ICounterTimingCallback {
     protected _interval: number = 300000;
     protected _resetTimeout: number = 0;
     protected _cache: { [id: string]: Counter } = {};
@@ -89,14 +89,14 @@ export abstract class CachedCounters implements ICounters, IReconfigurable, ITim
 
     /**
 	 * Begins measurement of execution time interval.
-	 * It returns [[Timing]] object which has to be called at
-	 * [[Timing.endTiming]] to end the measurement and update the counter.
+	 * It returns [[CounterTiming]] object which has to be called at
+	 * [[CounterTiming.endTiming]] to end the measurement and update the counter.
 	 * 
 	 * @param name 	a counter name of Interval type.
-	 * @returns a [[Timing]] callback object to end timing.
+	 * @returns a [[CounterTiming]] callback object to end timing.
      */
-    public beginTiming(name: string): Timing {
-        return new Timing(name, this);
+    public beginTiming(name: string): CounterTiming {
+        return new CounterTiming(name, this);
     }
 
     /**
@@ -205,7 +205,7 @@ export abstract class CachedCounters implements ICounters, IReconfigurable, ITim
      * @param name      a counter name
      * @param elapsed   execution elapsed time in milliseconds to update the counter.
      * 
-     * @see [[Timing.endTiming]]
+     * @see [[CounterTiming.endTiming]]
      */
     public endTiming(name: string, elapsed: number): void {
         let counter: Counter = this.get(name, CounterType.Interval);