+++ test/count/NullCounters.test.ts
@@ -1,5 +1,5 @@
 import { NullCounters } from '../../src/count/NullCounters';
-import { Timing } from '../../src/count/Timing';
+import { CounterTiming } from '../../src/count/CounterTiming';
 
 suite('NullCounters', ()=> {
 
@@ -14,7 +14,7 @@ suite('NullCounters', ()=> {
     test('Measure Elapsed Time', () => {
         let  counters: NullCounters = new NullCounters();
         
-        let timer: Timing = counters.beginTiming("Test.Elapsed");
+        let timer: CounterTiming = counters.beginTiming("Test.Elapsed");
         timer.endTiming();
     });