+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Component definitions for Node.js / ES2017 Changelog
 
+## <a name="1.2.0"></a> 1.2.0 (2021-04-09) 
+
+### Features
+* **trace** Added NullTracer class
+* **trace** Added LogTracer class
+* **trace** Added CachedTracer class
+* **trace** Added CompositeTracer class
+* Added tracer to Component class
+
 ## <a name="1.1.0"></a> 1.1.0 (2021-03-26) 
 
 ### Features